+++ src/Controller/VideosCategoriesController.php
@@ -55,11 +55,11 @@ class VideosCategoriesController extends AppController
     public function view($category = null)
     {
         //The category can be passed as query string, from a widget
-        if ($this->request->query('q')) {
-            return $this->redirect([$this->request->query('q')]);
+        if ($this->request->getQuery('q')) {
+            return $this->redirect([$this->request->getQuery('q')]);
         }
 
-        $page = $this->request->query('page') ? $this->request->query('page') : 1;
+        $page = $this->request->getQuery('page', 1);
 
         //Sets the cache name
         $cache = sprintf('index_category_%s_limit_%s_page_%s', md5($category), $this->paginate['limit'], $page);
@@ -83,7 +83,7 @@ class VideosCategoriesController extends AppController
                     },
                 ])
                 ->where(['Categories.slug' => $category, 'is_spot' => false])
-                ->order([sprintf('%s.created', $this->VideosCategories->Videos->alias()) => 'DESC']);
+                ->order([sprintf('%s.created', $this->VideosCategories->Videos->getAlias()) => 'DESC']);
 
             if ($query->isEmpty()) {
                 throw new RecordNotFoundException(__d('me_cms', 'Record not found'));
@@ -94,11 +94,11 @@ class VideosCategoriesController extends AppController
             //Writes on cache
             Cache::writeMany([
                 $cache => $videos,
-                sprintf('%s_paging', $cache) => $this->request->param('paging'),
+                sprintf('%s_paging', $cache) => $this->request->getParam('paging'),
             ], $this->VideosCategories->cache);
         //Else, sets the paging parameter
         } else {
-            $this->request->params['paging'] = $paging;
+            $this->request = $this->request->withParam('paging', $paging);
         }
 
         $this->set(am(['category' => $videos[0]->category], compact('videos')));