+++ src/Model/Table/VideosTable.php
@@ -22,12 +22,19 @@
  */
 namespace MeCmsYoutube\Model\Table;
 
+use ArrayObject;
 use Cake\Cache\Cache;
+use Cake\Datasource\EntityInterface;
+use Cake\Event\Event;
 use Cake\I18n\Time;
+use Cake\ORM\Entity;
+use Cake\ORM\Locator\LocatorAwareTrait;
 use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use MeCmsYoutube\Utility\Youtube;
 use MeCms\Model\Table\AppTable;
+use MeCms\Model\Table\Traits\IsOwnedByTrait;
+use MeCms\Model\Table\Traits\NextToBePublishedTrait;
 
 /**
  * Videos model
@@ -36,6 +43,10 @@ use MeCms\Model\Table\AppTable;
  */
 class VideosTable extends AppTable
 {
+    use IsOwnedByTrait;
+    use LocatorAwareTrait;
+    use NextToBePublishedTrait;
+
     /**
      * Name of the configuration to use for this table
      * @var string
@@ -60,13 +71,13 @@ class VideosTable extends AppTable
      * @param \ArrayObject $options Options
      * @return void
      * @uses MeCms\Model\Table\AppTable::afterDelete()
-     * @uses MeCms\Model\Table\AppTable::setNextToBePublished()
+     * @uses MeCms\Model\Table\Traits\NextToBePublishedTrait::setNextToBePublished()
      */
-    public function afterDelete(\Cake\Event\Event $event, \Cake\ORM\Entity $entity, \ArrayObject $options)
+    public function afterDelete(Event $event, Entity $entity, ArrayObject $options)
     {
         parent::afterDelete($event, $entity, $options);
 
-        //Sets the next video to be published
+        //Sets the next record to be published
         $this->setNextToBePublished();
     }
 
@@ -77,13 +88,13 @@ class VideosTable extends AppTable
      * @param \ArrayObject $options Options
      * @return void
      * @uses MeCms\Model\Table\AppTable::afterSave()
-     * @uses MeCms\Model\Table\AppTable::setNextToBePublished()
+     * @uses MeCms\Model\Table\Traits\NextToBePublishedTrait::setNextToBePublished()
      */
-    public function afterSave(\Cake\Event\Event $event, \Cake\ORM\Entity $entity, \ArrayObject $options)
+    public function afterSave(Event $event, Entity $entity, ArrayObject $options)
     {
         parent::afterSave($event, $entity, $options);
 
-        //Sets the next video to be published
+        //Sets the next record to be published
         $this->setNextToBePublished();
     }
 
@@ -91,12 +102,12 @@ class VideosTable extends AppTable
      * Called before each entity is saved.
      * Stopping this event will abort the save operation.
      * @param \Cake\Event\Event $event Event
-     * @param \Cake\ORM\Entity $entity Entity
+     * @param \Cake\Datasource\EntityInterface $entity Entity
      * @param \ArrayObject $options Options
      * @return bool
      * @uses _getInfo()
      */
-    public function beforeSave(\Cake\Event\Event $event, \Cake\ORM\Entity $entity, \ArrayObject $options)
+    public function beforeSave(Event $event, EntityInterface $entity, ArrayObject $options)
     {
         if ((empty($entity->seconds) || empty($entity->duration)) && !empty($entity->youtube_id)) {
             $info = $this->_getInfo($entity->youtube_id);
@@ -131,8 +142,8 @@ class VideosTable extends AppTable
      * @param string $type The type of query to perform
      * @param array|ArrayAccess $options An array that will be passed to Query::applyOptions()
      * @return Cake\ORM\Query The query builder
-     * @uses MeCms\Model\Table\AppTable::getNextToBePublished()
-     * @uses MeCms\Model\Table\AppTable::setNextToBePublished()
+     * @uses MeCms\Model\Table\Traits\NextToBePublishedTrait::getNextToBePublished()
+     * @uses MeCms\Model\Table\Traits\NextToBePublishedTrait::setNextToBePublished()
      */
     public function find($type = 'all', $options = [])
     {
@@ -159,9 +170,9 @@ class VideosTable extends AppTable
     public function findActive(Query $query, array $options)
     {
         $query->where([
-            sprintf('%s.active', $this->alias()) => true,
-            sprintf('%s.is_spot', $this->alias()) => false,
-            sprintf('%s.created <=', $this->alias()) => new Time,
+            sprintf('%s.active', $this->getAlias()) => true,
+            sprintf('%s.is_spot', $this->getAlias()) => false,
+            sprintf('%s.created <=', $this->getAlias()) => new Time,
         ]);
 
         return $query;
@@ -178,9 +189,9 @@ class VideosTable extends AppTable
         return $this->find()
             ->select('youtube_id')
             ->where([
-                sprintf('%s.active', $this->alias()) => true,
-                sprintf('%s.is_spot', $this->alias()) => true,
-                sprintf('%s.created <=', $this->alias()) => new Time,
+                sprintf('%s.active', $this->getAlias()) => true,
+                sprintf('%s.is_spot', $this->getAlias()) => true,
+                sprintf('%s.created <=', $this->getAlias()) => new Time,
             ])
             ->cache('all_spots', $this->cache)
             ->sample($limit)
@@ -196,20 +207,18 @@ class VideosTable extends AppTable
     {
         parent::initialize($config);
 
-        $this->table('youtube_videos');
-        $this->displayField('title');
-        $this->primaryKey('id');
+        $this->setTable('youtube_videos');
+        $this->setDisplayField('title');
+        $this->setPrimaryKey('id');
 
-        $this->belongsTo('Categories', [
-            'foreignKey' => 'category_id',
-            'joinType' => 'INNER',
-            'className' => 'MeCmsYoutube.VideosCategories',
-        ]);
-        $this->belongsTo('Users', [
-            'foreignKey' => 'user_id',
-            'joinType' => 'INNER',
-            'className' => 'MeCms.Users',
-        ]);
+        $this->belongsTo('Categories', ['className' => 'MeCmsYoutube.VideosCategories'])
+            ->setForeignKey('category_id')
+            ->setJoinType('INNER')
+            ->setTarget($this->tableLocator()->get('MeCmsYoutube.VideosCategories'));
+
+        $this->belongsTo('Users', ['className' => 'MeCms.Users'])
+            ->setForeignKey('user_id')
+            ->setJoinType('INNER');
 
         $this->addBehavior('Timestamp');
         $this->addBehavior('CounterCache', ['Categories' => ['video_count']]);
@@ -220,7 +229,7 @@ class VideosTable extends AppTable
     /**
      * Build query from filter data
      * @param Query $query Query object
-     * @param array $data Filter data ($this->request->query)
+     * @param array $data Filter data ($this->request->getQuery())
      * @return Query $query Query object
      * @uses \MeCms\Model\Table\AppTable::queryFromFilter()
      */
@@ -230,7 +239,7 @@ class VideosTable extends AppTable
 
         //"Is spot?" field
         if (!empty($data['spot']) && $data['spot']) {
-            $query->where([sprintf('%s.is_spot', $this->alias()) => true]);
+            $query->where([sprintf('%s.is_spot', $this->getAlias()) => true]);
         }
 
         return $query;